Type changed from integer to float (order inventory release qty field)

+++ server/graphql/types/order-inventory/new-order-inventory.ts
@@ -11,7 +11,7 @@ export const NewOrderInventory = gql`
     productName: String
     batchId: String
     packingType: String
-    releaseQty: Int
+    releaseQty: Float
     releaseWeight: Float
     remark: String
     issue: String
